Move Wakelock to TimerService
Because the TimerService is doing the work, the Wakelock should rather
be there instead of the GUI thread.
